refactor: assorted scope and symbol name improvements.
* lib/typecheck/init.lua: Use `pattern` consistently in symbol names
rather than a mixture of `pat`, `patt` and `match`.
Remove some unused function parameters and adjust callers accordingly.
Denote unused variables as `_` consistently.
(argscheck): Rename from this...
(checktypes): ...to this, to make it clearer that this is a non-API
function for checking annotated function arguments against type
predicate functions.  Encapsulate into a setup function.
(argcheck): Encapsulate into a setup function.
(normalize_argscheck): Rename from this...
(argscheck): ...to this, instead of renaming in the export table where
it was easily conflated with the other previous argscheck function.
Encapsulate into a setup function.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>